Week_5/requirements.txt
snowflake-connector-python
pandas
python-dotenv
streamlit
altair
